+++ NEWS.md
@@ -4,6 +4,7 @@
 * Modify the post processing of `MHT01` to allow multiple `row_split_var`.
 * Improve the `report_null` method to facilitate the creation of null reports
 * Export the `std_postprocessing` function to simplify post processing logic.
+* `MNG01` line type can now be controlled with the `line_type` argument.
 
 # chevron 0.2.6